requirements.txt
aiogram==2.5
langchain==0.0.225
python-dotenv